requirements.txt
webdriver-manager
selenium
hugchat
matplotlib
seaborn
python-dotenv